+++ src/main/java/com/example/assignment/admin/dto/AdminSettlementPageRes.java
@@ -1,6 +1,7 @@
-package com.example.assignment.domain.dto.response;
+package com.example.assignment.admin.dto;
 
-import com.example.assignment.domain.dto.PageResponse;
+import com.example.assignment.common.dto.PageResponse;
+import com.example.assignment.creator.dto.response.CourseRes;
 import java.text.NumberFormat;
 import java.util.List;
 import java.util.Locale;
@@ -8,6 +9,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.data.domain.Page;
 
 @Getter
 @Builder
@@ -22,16 +24,25 @@ public class AdminSettlementPageRes {
   private String totalSettlementAmount;
   private PageResponse<AdminSettlementRes> pageResponse;
 
-  public static AdminSettlementPageRes toResponse(List<AdminSettlementRes> settlements, int page) {
+  public static AdminSettlementPageRes toResponse(Page<AdminSettlementRes> settlementPage) {
+    List<AdminSettlementRes> settlements = settlementPage.getContent();
+
     long totalAmount      = settlements.stream().mapToLong(AdminSettlementRes::getRawTotalAmount).sum();
     long totalCommission  = settlements.stream().mapToLong(AdminSettlementRes::getRawCommission).sum();
     long totalSettlement  = settlements.stream().mapToLong(AdminSettlementRes::getRawSettlementAmount).sum();
+
+    PageResponse<AdminSettlementRes> response = new PageResponse<>(
+        settlementPage.getNumber() + 1,
+        settlementPage.getTotalPages(),
+        settlementPage.getTotalElements(),
+        settlements
+    );
 
     return AdminSettlementPageRes.builder()
         .totalAmount(FORMAT.format(totalAmount) + "원")
         .totalCommission(FORMAT.format(totalCommission) + "원")
         .totalSettlementAmount(FORMAT.format(totalSettlement) + "원")
-        .pageResponse(PageResponse.pagination(settlements, page))
+        .pageResponse(response)
         .build();
   }
 }
